# What
This PR starts sending last read chapter details in the manga object. The provision to send this was already there but was not utilized for some reason(perhaps performance reasons, but that's just speculation)

# Why
This will enable us to implement the sort by last read manga feature

# Performance report for API `api/v1/category/{categoryId}`
JS code used for benchmarking:
```javascript
var url = "http://127.0.0.1:4567/api/v1/category/0"
const numRequests = 10
let totalFetchTime = 0;

for (let i = 0; i < numRequests; i++) {
    const startTime = performance.now();
    fetch(url)
        .then(response => {
            const endTime = performance.now();
            const fetchTime = endTime - startTime;
            console.log(`Request ${i+1}: ${fetchTime} ms`);
            totalFetchTime += fetchTime;
        })
        .catch(error => console.error(error));
}

const averageFetchTime = totalFetchTime / numRequests;
console.log(`Average fetch time: ${averageFetchTime} ms`);
```
Total Number of Rows in Manga table: `909`

Total Number of Manga in Library: `93`

Total Number of Rows in Chapter table: `18760`

Total Number of Chapters for Mangas in Library: `18345`

## TIMINGS
### Optimized Query after Last Read Chapter change
Average Request time over 10 requests ran 3 times: `392.173333333 ms`
### Non Optimized query after Last Read Chapter change
Average Request time over 10 requests ran 3 times: `430.143333333 ms`
### Optimized Query without Last Read Chapter change
Average Request time over 10 requests ran 3 times: `286.22 ms`
### Optimized Query with new column instead of Chapter object
Average Request time over 10 requests ran 3 times: `304.553333333 ms`
### Baseline(master branch)
Average Request time over 10 requests ran 3 times: `315.47 ms`

Server Hardware details:
```
CPU: AMD Ryzen 9 5900X 12-Core Processor
RAM: 32GB DDR4 3200 MT/s
STORAGE: Samsung SSD 970 EVO Plus 500GB
```