@AriaMoradi 2023-02-20
I hope h2 knows how to optimize queries, oof we are now hitting chapters multiple times, at least it's in the number of hundreds to thousands so hopefully it won't become slower and slower as we make more high level data available...